SDESK-5968 UPDATE: Moved toClientFormat function to utils.
